# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sohati — Egyptian Medicine Verification Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Medicine {
  id: string;
  /** @nullable */
  tradeNameEn?: string | null;
  /** @nullable */
  tradeNameAr?: string | null;
  /** @nullable */
  genericName?: string | null;
  /** @nullable */
  applicantName?: string | null;
  /** @nullable */
  dosageForm?: string | null;
  /** @nullable */
  registrationNumber?: string | null;
  /** @nullable */
  barcode?: string | null;
  isVerified: boolean;
  /** @nullable */
  source?: string | null;
  createdAt: string;
}

export interface BarcodeResult {
  found: boolean;
  medicine?: Medicine;
}

export interface BarcodeNotFound {
  found: boolean;
  message: string;
}

export interface BarcodeMappingInput {
  barcode: string;
  medicineId: string;
}

export interface ReportInput {
  medicineName: string;
  reportReason: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  governorate?: string | null;
  /** @nullable */
  pharmacyName?: string | null;
  /** @nullable */
  contactNumber?: string | null;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatInput {
  messages: ChatMessage[];
  /** @nullable */
  medicineContext?: string | null;
}

export interface ChatResponse {
  reply: string;
}

export interface ImageAnalysisInput {
  imageBase64: string;
  mimeType: string;
}

export interface ImageAnalysisResult {
  /** @nullable */
  drugName?: string | null;
  /** @nullable */
  confidence?: string | null;
}

export interface PlatformStats {
  medicineCount: number;
  reportCount: number;
  userCount: number;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface VerifyMedicineInput {
  disease: string;
  tradeName: string;
}

export interface VerifyMedicineResult {
  isForDisease: boolean;
  explanation: string;
  alternatives?: string[];
}

export interface ErrorResponse {
  error: string;
}

export type SearchMedicinesParams = {
q: string;
type?: SearchMedicinesType;
};

export type SearchMedicinesType = typeof SearchMedicinesType[keyof typeof SearchMedicinesType];


export const SearchMedicinesType = {
  name: 'name',
  barcode: 'barcode',
} as const;