# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sohati — Egyptian Medicine Verification Platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: medicines
    description: Medicine search and verification
  - name: reports
    description: Suspicious medicine reports
  - name: assistant
    description: AI pharmacist assistant
  - name: stats
    description: Platform statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /medicines/search:
    get:
      operationId: searchMedicines
      tags: [medicines]
      summary: Search medicines by name or barcode
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [name, barcode]
            default: name
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Medicine"

  /medicines/barcode/{barcode}:
    get:
      operationId: getMedicineByBarcode
      tags: [medicines]
      summary: Look up medicine by barcode
      parameters:
        - name: barcode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Medicine found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarcodeResult"
        "404":
          description: Barcode not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarcodeNotFound"

  /medicines/{id}:
    get:
      operationId: getMedicine
      tags: [medicines]
      summary: Get full medicine details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Medicine details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Medicine"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /medicines/barcode-map:
    post:
      operationId: submitBarcodeMapping
      tags: [medicines]
      summary: Submit a barcode-to-medicine mapping (crowdsourced)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BarcodeMappingInput"
      responses:
        "201":
          description: Mapping saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /reports:
    post:
      operationId: createReport
      tags: [reports]
      summary: Submit a suspicious medicine report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /assistant/chat:
    post:
      operationId: chatWithAssistant
      tags: [assistant]
      summary: Chat with the AI pharmacist assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "200":
          description: Assistant response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /assistant/verify-medicine:
    post:
      operationId: verifyMedicine
      tags: [assistant]
      summary: Verify if a medicine trade name is truly for a given disease
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyMedicineInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyMedicineResult"

  /assistant/analyze-image:
    post:
      operationId: analyzeImage
      tags: [assistant]
      summary: Extract drug name from medicine box photo (base64)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageAnalysisInput"
      responses:
        "200":
          description: Extracted drug name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageAnalysisResult"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get platform statistics (medicine count, reports, users)
      responses:
        "200":
          description: Platform stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Medicine:
      type: object
      properties:
        id:
          type: string
        tradeNameEn:
          type: ["string", "null"]
        tradeNameAr:
          type: ["string", "null"]
        genericName:
          type: ["string", "null"]
        applicantName:
          type: ["string", "null"]
        dosageForm:
          type: ["string", "null"]
        registrationNumber:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        isVerified:
          type: boolean
        source:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - isVerified
        - createdAt

    BarcodeResult:
      type: object
      properties:
        found:
          type: boolean
        medicine:
          $ref: "#/components/schemas/Medicine"
      required:
        - found

    BarcodeNotFound:
      type: object
      properties:
        found:
          type: boolean
        message:
          type: string
      required:
        - found
        - message

    BarcodeMappingInput:
      type: object
      properties:
        barcode:
          type: string
        medicineId:
          type: string
      required:
        - barcode
        - medicineId

    ReportInput:
      type: object
      properties:
        medicineName:
          type: string
        reportReason:
          type: string
        description:
          type: ["string", "null"]
        governorate:
          type: ["string", "null"]
        pharmacyName:
          type: ["string", "null"]
        contactNumber:
          type: ["string", "null"]
      required:
        - medicineName
        - reportReason

    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content

    ChatInput:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
        medicineContext:
          type: ["string", "null"]
      required:
        - messages

    ChatResponse:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply

    ImageAnalysisInput:
      type: object
      properties:
        imageBase64:
          type: string
        mimeType:
          type: string
      required:
        - imageBase64
        - mimeType

    ImageAnalysisResult:
      type: object
      properties:
        drugName:
          type: ["string", "null"]
        confidence:
          type: ["string", "null"]
      required: []

    PlatformStats:
      type: object
      properties:
        medicineCount:
          type: integer
        reportCount:
          type: integer
        userCount:
          type: integer
      required:
        - medicineCount
        - reportCount
        - userCount

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success

    VerifyMedicineInput:
      type: object
      properties:
        disease:
          type: string
        tradeName:
          type: string
      required:
        - disease
        - tradeName

    VerifyMedicineResult:
      type: object
      properties:
        isForDisease:
          type: boolean
        explanation:
          type: string
        alternatives:
          type: array
          items:
            type: string
      required:
        - isForDisease
        - explanation

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error