# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sohati — Egyptian Medicine Verification Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatInput {
  messages: ChatMessage[];
  /** @nullable */
  medicineContext?: string | null;
}