# lib/api-zod/src/generated/types/medicine.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sohati — Egyptian Medicine Verification Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Medicine {
  id: string;
  /** @nullable */
  tradeNameEn?: string | null;
  /** @nullable */
  tradeNameAr?: string | null;
  /** @nullable */
  genericName?: string | null;
  /** @nullable */
  applicantName?: string | null;
  /** @nullable */
  dosageForm?: string | null;
  /** @nullable */
  registrationNumber?: string | null;
  /** @nullable */
  barcode?: string | null;
  isVerified: boolean;
  /** @nullable */
  source?: string | null;
  createdAt: string;
}