# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sohati — Egyptian Medicine Verification Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportInput {
  medicineName: string;
  reportReason: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  governorate?: string | null;
  /** @nullable */
  pharmacyName?: string | null;
  /** @nullable */
  contactNumber?: string | null;
}