# lib/api-zod/src/generated/types/verifyMedicineInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sohati — Egyptian Medicine Verification Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface VerifyMedicineInput {
  disease: string;
  tradeName: string;
}